chore: Refractor create org form components
